reformats registrarCliente
