+++ ClubDeportivo/Entidades/Socio.cs
@@ -10,22 +10,5 @@ namespace ClubDeportivo
             this.Cuota = cuota;
         }
 
-        // Si bien es igual al metodo de NoSocio, en el futuro pueden variar
-        public override int? RegistrarCliente(Cliente cliente)
-        {
-            try
-            {
-                Clientes clientesDb = new();
-                var id = clientesDb.RegistrarCliente(cliente);
-                if (id == 0) throw new Exception("No se pudo registrar el cliente");
-                return id;
-            }
-            catch (Exception e)
-            {
-                throw new Exception(e.Message);
-            }
-        }
-
-
     }
 }
